+++ pczt/CHANGELOG.md
@@ -10,6 +10,13 @@ workspace.
 
 ## [Unreleased]
 
+### Added
+- `pczt::roles::signer`:
+  - `Signer::sighash`
+  - `Signer::append_transparent_signature`
+  - `Signer::apply_sapling_signature`
+  - `Signer::apply_orchard_signature`
+
 ## [0.5.0] - 2025-11-05
 
 ### Changed